Add files via upload
Q2 . Given an array arr, return the second largest element from an array. If the second largest element doesn't exist then return -1.
Note: The second largest element should not be equal to the first largest.

Input: arr = [12, 35, 1, 10, 34, 1]
Output: 34
Explanation: The largest element of the array is 35 and the second largest element is 34.